scaffolding-chef: secure ssl default
We should use the safe and secure default of `:verify_peer` for the
`ssl_verify_mode` in the configuration generated by the Chef
scaffolding. This is in line with the default within Chef itself. End
users can override this with `:verify_none` by changing the
configuration passed into Habitat if they wish.

Signed-off-by: Joshua Timberman <joshua@chef.io>